refactor: break out session queries
